Source files

+++ app/remove_match.php
@@ -1,25 +1,37 @@
 <?php
 $pageTitle = "Usuwanie meczu";
 include __DIR__ . '/global/section/header.php';
-
 require __DIR__ . '/global/global-db/db_config.php';
 
-if ($_SERVER["REQUEST_METHOD"] == "POST" && isset($_POST['match_id'])) {
-    $matchId = $_POST['match_id'];
+session_start();
+$isLogged = $_SESSION['isLogged'];
 
-    // Zapytanie SQL do usunięcia meczu
-    $sqlRemoveMatch = "DELETE FROM matches WHERE id = '$matchId'";
-    $resultRemoveMatch = $conn->query($sqlRemoveMatch);
+if ($isLogged == true) {
+    if ($_SERVER["REQUEST_METHOD"] == "POST" && isset($_POST['match_id'])) {
+        $matchId = $_POST['match_id'];
 
-    if ($resultRemoveMatch) {
-        echo "Mecz o ID $matchId został usunięty.";
+        // Zapytanie SQL do usunięcia meczu
+        $sqlRemoveMatch = "DELETE FROM matches WHERE id = '$matchId'";
+        $resultRemoveMatch = $conn->query($sqlRemoveMatch);
+
+        if ($resultRemoveMatch) {
+            echo "Mecz o ID $matchId został usunięty.";
+        } else {
+            echo "Błąd podczas usuwania meczu: " . $conn->error;
+        }
     } else {
-        echo "Błąd podczas usuwania meczu: " . $conn->error;
+        echo "Nieprawidłowe żądanie.";
     }
-} else {
-    echo "Nieprawidłowe żądanie.";
-}
 
-// Zamknięcie połączenia
-$conn->close();
+    // Zamknięcie połączenia
+    $conn->close();
+} else {
+    include __DIR__ . '/global/section/not-available.html';
+}
 ?>
+
+<?php
+include __DIR__ . '/global/section/footer.php';
+?>
+
+</html>
